chore: soft net isolation for mac
